+++ journal/journal_test.go
@@ -11,11 +11,16 @@ import (
 func TestJounalCanBeCalled(t *testing.T) {
 	must, wont := hamlet.Specifications(t)
 
-	wont.Nil(must)
-
 	must.Equal("foo bar", journal.Unify("  foo  \t  \r\n   bar  "))
 
 	common.ControllerType = "unittest"
 
-	must.Nil(journal.Post("unittest", "journal", "from journal/journal_test.go"))
+	must.Nil(journal.Post("unittest", "journal-1", "from journal/journal_test.go"))
+	events, err := journal.Events()
+	must.Nil(err)
+	wont.Nil(events)
+	must.True(len(events) > 0)
+	must.Nil(journal.Post("unittest", "journal-2", "from journal/journal_test.go"))
+	second, err := journal.Events()
+	must.True(len(second) > len(events))
 }